report-service/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
aiokafka
pydantic
pydantic-settings